Update fr.md
Removed the `Graven - Développement` channel from the French avoid list.

Effectively, the series that could, in the past, be problematic with bad practices... are now, for the most, marked as deprecated. I'll take the example of the old Web series : this series has, 1 month ago, placed into a playlist that explicitly marks the content as deprecated.
For the other mistakes on the series, even the websites you give as references can contains content that is not recommended nowadays, but it doesn't make the courses bad.

Furthermore, I want to note that the pull request that added this line to the file has been made after a conflict between the author of the pull request and the concerned YouTuber. 

So, can you consider this pull request and do not reject it with only one approbation.